+++ HierarchyControl/HierarchyControl/ControlManifest.Input.xml
@@ -1,61 +1,59 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest>
-  <control namespace="Carfup" constructor="HierarchyControl" version="1.0.2"
-    display-name-key="Carfup.HierarchyControl"
-    description-key="Control allowing you to visualize the hierarchy between records."
-    control-type="virtual">
-    <!--external-service-usage
-    node declares whether this 3rd party PCF control is using external service or not, if yes, this
-    control will be considered as premium and please also add the external domain it is using.
-    If it is not using any external service, please set the enabled="false" and DO NOT add any domain
-    below. The "enabled" will be false by default.
-    Example1:
-      <external-service-usage enabled="true">
-        <domain>www.Microsoft.com</domain>
-      </external-service-usage>
-    Example2:
-      <external-service-usage enabled="false">
-      </external-service-usage>
-    -->
-    <external-service-usage enabled="false">
-      <!--UNCOMMENT
-      TO ADD EXTERNAL DOMAINS
-      <domain></domain>
-      <domain></domain>
-      -->
-    </external-service-usage>
-    <!-- property node identifies a specific, configurable piece of data that the control expects
-    from CDS -->
-    <property name="hostingField" display-name-key="Hosting Field"
-      description-key="Field where to attach the control" of-type="SingleLine.Text" usage="bound"
-      required="true" />
-    <property name="JsonMapping" display-name-key="JSON Mapping"
-      description-key="JSON Mapping to define the values to retrieve" of-type="Multiple"
-      usage="input" required="true" />
-    <!--
-      Property node's of-type attribute can be of-type-group attribute.
-      Example:
-      <type-group name="numbers">
-        <type>Whole.None</type>
-        <type>Currency</type>
-        <type>FP</type>
-        <type>Decimal</type>
-      </type-group>
-      <property name="sampleProperty" display-name-key="Property_Display_Key"
-    description-key="Property_Desc_Key" of-type-group="numbers" usage="bound" required="true" />
-    -->
-    <resources>
-      <code path="index.ts" order="1" />
-      <platform-library name="React" version="16.14.0" />
-      <platform-library name="Fluent" version="9.46.2" />
-      <!-- UNCOMMENT TO ADD MORE RESOURCES
-      <css path="css/HierarchyControlF.css" order="1" />
-      <resx path="strings/HierarchyControlF.1033.resx" version="1.0.0" />
-      -->
-    </resources>
-    <feature-usage>
-      <uses-feature name="Utility" required="true" />
-      <uses-feature name="WebAPI" required="true" />
-    </feature-usage>
-  </control>
+	<control namespace="Carfup" constructor="HierarchyControl" version="1.1.0" display-name-key="Carfup.HierarchyControl" description-key="Control allowing you to visualize the hierarchy between records." control-type="virtual">
+		<!--
+		external-service-usage
+		node declares whether this 3rd party PCF control is using external service or not, if yes, this
+		control will be considered as premium and please also add the external domain it is using.
+		If it is not using any external service, please set the enabled="false" and DO NOT add any domain
+		below. The "enabled" will be false by default.
+		Example1:
+		<external-service-usage enabled="true">
+		<domain>www.Microsoft.com</domain>
+		</external-service-usage>
+		Example2:
+		<external-service-usage enabled="false">
+		</external-service-usage>
+		-->
+		<external-service-usage enabled="false">
+			<!--
+			UNCOMMENT
+			TO ADD EXTERNAL DOMAINS
+			<domain></domain>
+			<domain></domain>
+			-->
+		</external-service-usage>
+		<!--
+		property node identifies a specific, configurable piece of data that the control expects
+		from CDS
+		-->
+		<property name="hostingField" display-name-key="Hosting Field" description-key="Field where to attach the control" of-type="SingleLine.Text" usage="bound" required="true" />
+		<property name="JsonMapping" display-name-key="JSON Mapping" description-key="JSON Mapping to define the values to retrieve" of-type="Multiple" usage="input" required="true" />
+		<!--
+		Property node's of-type attribute can be of-type-group attribute.
+		Example:
+		<type-group name="numbers">
+		<type>Whole.None</type>
+		<type>Currency</type>
+		<type>FP</type>
+		<type>Decimal</type>
+		</type-group>
+		<property name="sampleProperty" display-name-key="Property_Display_Key"
+		description-key="Property_Desc_Key" of-type-group="numbers" usage="bound" required="true" />
+		-->
+		<resources>
+			<code path="index.ts" order="1" />
+			<platform-library name="React" version="16.14.0" />
+			<platform-library name="Fluent" version="9.46.2" />
+			<!--
+			UNCOMMENT TO ADD MORE RESOURCES
+			<css path="css/HierarchyControlF.css" order="1" />
+			<resx path="strings/HierarchyControlF.1033.resx" version="1.0.0" />
+			-->
+		</resources>
+		<feature-usage>
+			<uses-feature name="Utility" required="true" />
+			<uses-feature name="WebAPI" required="true" />
+		</feature-usage>
+	</control>
 </manifest>
